mynews/requirements.txt
Flask>=2.0
SQLAlchemy>=1.3
Flask-SQLAlchemy>=2.5
APScheduler>=3.6
feedparser>=6.0
requests>=2.0
python-dotenv>=0.20